Update README.md
I found that older versions of this readme were directing users to a possible malware site.   I've updated the file by adding footnotes with the suspect malware and I also added sources for MinGW that seem to be less of a risk.   It would be helpful if the owner of this project could double check my work is correct and that the new sources have files that download, install and make the project work with no dependency problems as of September 2023.  I will be doing that download and verification later this week but I can not do it at the moment.